fix: remove redundant inner uintptr_t cast flagged by clang-tidy on 64-bit
refs: RUM-17329

start_addr and end_addr are already declared as uintptr_t, so the inner
static_cast<uintptr_t> was a no-op. On 64-bit Linux uintptr_t and uint64_t
share the same underlying type (unsigned long), causing clang-tidy's
readability-redundant-casting check to fire as an error. The outer
static_cast<uint64_t> alone is sufficient: it widens on 32-bit and is a
named-type cast on 64-bit, which clang-tidy does not flag.

+++ src/datadog/impl/crash_reporting/handlers/inprocess/crash_handler_inprocess_posix.cpp
@@ -583,8 +583,8 @@ static void write_modules(int fd) {
       // included in our crash report: write that information to the output file
       WriteCrashReportModule(
           fd,
-          static_cast<uint64_t>(static_cast<uintptr_t>(start_addr)),
-          static_cast<uint64_t>(static_cast<uintptr_t>(end_addr)),
+          static_cast<uint64_t>(start_addr),
+          static_cast<uint64_t>(end_addr),
           pathname,
           build_id
       );
